1. Sapphire_App PLCService: Reactor Target값 publish 버그 수정 2. Sapphire_App Manual Batch: Ramp Time 유효한 값 입력 시 Load 버튼 활성화 안되는 문제 해결 3. Sapphire_App Monitoring Meter: 출력값을 정수값에서 소수점 값 형식으로 변경

+++ SapphireXR_App/ViewModels/PresentValueMonitor/MonitoringMeterViewModel.cs
@@ -4,7 +4,7 @@
     {
         protected override void updatePresentValue(float value)
         {
-            PresentValue = ((int)value).ToString();
+            PresentValue = (value).ToString();
         }
     }
 }
